Attempt to use `libcpp.timed_mutex` for locks
This is actually less performant than the current method
(as the following commits will show).

Obviously this is a naive attempt since the GIL must be released
before any blocking operations.

It *could* be improved by using `unique_lock` for ownership ident,
but that would require a bit more refactoring across the codebase.

Committing here for future reference.